requirements.txt
python-dotenv==1.0.1 # For reading environment variables stored in .env file
langchain  # For building applications with LLMs
langchain-community  # Community contributed packages
langchain-google-genai  # For Google Gemini LLM and embeddings
langchain-chroma  # For Chroma integration with LangChain
unstructured==0.14.4 # Document loading
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb  # Vector storage
google-generativeai  # For Google Gemini API
tiktoken==0.7.0  # For token counting

# install markdown depenendies with: `pip install "unstructured[md]"` after install the requirements file. Leave this line commented out. 
